♻️ 型定義の分離とMapper作成
## 概要
DBスキーマ型とドメインモデルを完全に分離し、Mapperクラスで相互変換を行う構造に変更。
infrastructure層の責務を明確化し、保守性と拡張性を向上。

## 詳細
- `infrastructure/supabase/schema.ts` 作成
  - Supabaseから取得する生のデータ型定義（`RecipeRow`）
  - DBカラム名そのまま（snake_case）
- `infrastructure/mappers/RecipeMapper.ts` 作成
  - `toDomain(row: RecipeRow): Recipe` - DB → ドメイン変換
  - `toRow(recipe: Recipe): Omit<RecipeRow, 'id' | 'created_at'>` - ドメイン → DB変換
  - repositories/から移動して適切な配置に変更
- `infrastructure/repositories/SupabaseRecipeRepository.ts` 修正
  - RecipeMapperを使用（private変換メソッド削除）
  - コード重複を解消
- `lib/supabase.ts` 修正
  - 型定義のre-export削除（supabaseクライアントのexportのみに責務を限定）
- `infrastructure/dto/RecipeDTO.ts` 削除
  - RecipeRowに統一

## 技術的詳細
- 型名の明確化: `RecipeDTO` → `RecipeRow`（DB型であることを明示）
- Mapperの配置: `repositories/` → `mappers/`（責務の明確化）
- 変換ロジックの統一: SupabaseRecipeRepositoryの変換ロジックをRecipeMapperに集約

## 受け入れ条件
- ✅ RecipeRow型が定義されている
- ✅ RecipeMapperが双方向変換を実装している
- ✅ SupabaseRecipeRepositoryがMapperを使用している
- ✅ lib/supabase.tsから型定義が削除されている
- ✅ 既存機能にデグレなし（単体テスト全て成功）
- ✅ TypeScriptビルドエラー: 0件
- ✅ ESLintエラー: 0件

## 関連Issue
#27
