+++ src/lib/supabase.ts
@@ -19,5 +19,3 @@ if (!supabaseUrl || !supabaseAnonKey) {
 }
 
 export const supabase = createClient(supabaseUrl, supabaseAnonKey);
-
-export type { RecipeDTO as Recipe } from '../infrastructure/dto/RecipeDTO';
