+++ supabase/.preview-trigger
@@ -42,3 +42,5 @@ Updated at: 2026-04-17 16:25:16
 Updated at: 2026-04-17 16:45:42
 
 Updated at: 2026-04-19 15:20:33
+
+Updated at: 2026-04-19 15:44:07
